+++ truffle-core/contracts/Utils.sol
@@ -38,7 +38,7 @@ contract Utils is EncodingMatchLog, AssetsView{
         bool[3] memory aligned1stSubst1stRedCardLastGame,
         uint8[3] memory generationGamesNonStopInjuryWeeks
     ) {
-        for (uint8 sk = 0; sk < N_SKILLS; sk++) skills[sk] = uint16(getSkill(encodedSkills, sk));
+        for (uint8 sk = 0; sk < N_SKILLS; sk++) skills[sk] = uint32(getSkill(encodedSkills, sk));
 
         dayOfBirth = uint16(getBirthDay(encodedSkills));
